Move load data function from main script to generate data script.
