+++ k-means/generate_clustering_data.py
@@ -7,7 +7,30 @@ Created on Mon Jun 29 14:55:33 2015
 
 import numpy as np
 import matplotlib.pyplot as plt
+from scipy import loadtxt
+from sklearn import datasets
 from sklearn import preprocessing
+
+
+def load_data(data_name='faithful'):
+    if data_name is 'faithful':
+        data = loadtxt('faithful.txt')
+        return preprocessing.scale(data)
+    elif data_name is 'generate_clustering_data':
+        data = generate_clustering_data.generate_clustering_data()
+        return preprocessing.scale(data)
+    elif data_name is 'generate_clustering_data_easy':
+        data = generate_clustering_data.generate_clustering_data_easy()
+        return preprocessing.scale(data)
+    elif data_name is 'generate_clustering_data_tilted':
+        data = generate_clustering_data.generate_clustering_data_tilted()
+        return preprocessing.scale(data)
+    elif data_name is 'digits':
+        data = datasets.load_digits().data
+        return preprocessing.scale(data)
+    elif data_name is 'iris':
+        data = datasets.load_iris().data
+        return preprocessing.scale(data)
 
 
 def generate_clustering_data():
